2021 12 05 pv fmx card 25 (#74)
 * Fixed residual risk logic in the RIPD template 
 * added a J2 placeholder for the iframe loading logic
 * added oauth0 JWT validation to the graph APIs
 * moved KPIs to a separate Resources file 
 * fixed issues with JWT Token Filter's fakeToken
