+++ config/index.html
@@ -4,8 +4,9 @@
     <meta charset="UTF-8">
     <title>Pontus Vision</title>
 </head>
-<body style="width: 100%; height: 100%; margin: 0px">
-<iframe id="pvIframe" style="width: 100%; height: 100%; margin: 0px" src="/grafana/"></iframe>
+<body style="width: 100%; height: 100%; margin: 0px; overflow: hidden">
+<iframe id="pvIframe" style="width: 100%; height: 100%; margin: 0px; overflow: hidden"
+        src="{{ iframe | default('/grafana/') }}"></iframe>
 <script src="/auth/js/keycloak.js"></script>
 <script>
   window.pvLoad = async () => {
