+++ src/test/java/com/pontusvision/gdpr/PVEndPointsTest.java
@@ -72,10 +72,10 @@ public class PVEndPointsTest extends AppTest {
       HttpGet getRequest = new HttpGet(URI.create("http://localhost:3001/home/param?name=Omar"));
 
 //      getRequest.setHeader("Content-Type", "application/json");
-      getRequest.setHeader("AUTHORIZATION", ": granted");
+      getRequest.setHeader("X-PV-NAME", ": granted");
 
       String res = IOUtils.toString(client.execute(getRequest).getEntity().getContent());
-      assertEquals("Hello, Omar AUTHORIZATION: granted", res, "message from ~/home/param endpoint");
+      assertEquals("Hello, Omar X-PV-NAME: granted", res, "message from ~/home/param endpoint");
 
     } catch (IOException e) {
       e.printStackTrace();
